move some more tests
